requirements.txt
# python==3.10
gym==0.26.2
h5py==3.10.0
joblib==1.3.2
matplotlib==3.8.1
numpy==1.24.3
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.3.2
scipy==1.11.3
setuptools==68.0.0
torch==1.13.1
